+++ changelog.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 This project attempts to adhere to [Semantic Versioning](http://semver.org/).
 
-## [4.6.0] - [unreleased]
+## [4.6.2] - [unreleased]
+### Changed
+- changed Iterables#rangeTo(int,int) and Iterables#rangeTo(int,int,int) to properly handle max and min integers
+- changed rangeTo().iterator().next() to throw NoSuchElementException when hasNext() returns false
+
+## [4.6.1] - [2018-03-12]
 ### Added
 - added `Pair#zip(Optional, Optional)`
 - added `Eithers#sequenceRight(Iterable, Collector)`